@@ -7,7 +7,7 @@
         "test": "jest",
         "build": "tsc",
         "watch": "tsc -w --preserveWatchOutput",
-        "lint": "tslint '*.ts'",
+        "lint": "eslint \"src/**/*.ts\"",
         "create-tables": "run-script-os",
         "create-tables:nix": "./__test__/environment/local/create_tables.sh",
         "create-tables:win32": "cd ./__test__/environment/local/ && @powershell -NoProfile -ExecutionPolicy Unrestricted -Command ./create_tables.ps1",
@@ -21,22 +21,27 @@
         "aws-embedded-metrics": "^2.0.4",
         "aws-sdk": "^2.437.0",
         "source-map-support": "^0.5.19",
-        "uuid": "^3.1.0"
+        "uuid": "^8.3.2"
     },
     "devDependencies": {
         "@tsconfig/node14": "^1.0.0",
         "@types/aws-lambda": "^8.10.72",
-        "@types/jest": "^26.0.23",
-        "@types/node": "^14.14.26",
+        "@types/jest": "^27.0.2",
+        "@types/node": "^16.9.4",
         "@types/uuid": "^8.3.0",
+        "@typescript-eslint/eslint-plugin": "^4.31.2",
+        "@typescript-eslint/eslint-plugin-tslint": "^4.31.2",
+        "@typescript-eslint/parser": "^4.31.2",
         "concurrently": "^6.3.0",
-        "jest": "^26.6.3",
+        "eslint": "^7.32.0",
+        "eslint-config-prettier": "^8.3.0",
+        "eslint-plugin-import": "^2.24.2",
+        "eslint-plugin-jsdoc": "^36.1.0",
+        "jest": "^27.2.1",
         "koekalenteri-shared": "file:../koekalenteri-shared",
         "run-script-os": "^1.1.6",
-        "ts-jest": "^26.5.6",
-        "ts-node": "^9.1.1",
-        "tslint": "^5.5.0",
-        "tslint-config-standard": "^6.0.1",
-        "typescript": "^4.1.5"
+        "ts-jest": "^27.0.5",
+        "ts-node": "^10.2.1",
+        "typescript": "^4.4.3"
     }
 }
