@@ -1,6 +1,8 @@
 import { createContext } from "react";
 import { EventStore } from "./EventStrore";
+import { JudgeStore } from "./JudgeStore";
 
 export const rootStoreContext = createContext({
-  eventStore: new EventStore()
+  eventStore: new EventStore(),
+  judgeStore: new JudgeStore()
 });
